+++ wire-ios/Wire-iOS/Sources/Helpers/ZMConversationMessage+Announcement.swift
@@ -46,7 +46,7 @@ extension ZMConversationMessage {
             return ConversationAnnouncement.Video.description(senderName)
         } else if isFile {
             return ConversationAnnouncement.File.description((filename ?? ""), senderName)
-        } else if isSystem, let cellDescription = ConversationSystemMessageCellDescription.cells(for: self, isCollapsed: true).first {
+        } else if isSystem, let cellDescription = ConversationSystemMessageCellDescription.cells(for: self).first {
             return cellDescription.cellAccessibilityLabel
         }
         return nil
